requirements.txt
discord.py
asyncpg
discord.py